Finally fixed ALL the function prototypes so that there are NO warnings when gamesman compiles.  YAY!      -Dom
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@1856 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/mcon.cc
@@ -28,8 +28,8 @@ EXTERNC
 
 POSITION gNumberOfPositions  =  0;
 POSITION gInitialPosition    =  0;
-POSITION gMinimalPosition    = 0 ;
-POSITION kBadPosition        = -1; /* This can never be the rep. of a position */
+POSITION gMinimalPosition    =  0;
+POSITION kBadPosition        = (unsigned long long)-1; /* This can never be the rep. of a position */
 
 STRING   kAuthorName         = "Brian Carnes";
 STRING   kGameName           = "Connections";
